+++ stdlib/LinearAlgebra/src/LinearAlgebra.jl
@@ -126,7 +126,6 @@ export
     pinv,
     qr,
     qrfact!,
-    qrfact,
     lq,
     lqfact!,
     rank,